refactor(versions): introduce Port + Adapter layering as #154 pilot
Establish the regulative Hexagonal structure for the versions domain that
the remaining seven sub-issues (#222-#228) will replicate. Adds:

- `app/versions/domain/ports.py` — `VersionRepository` runtime-checkable
  Protocol covering all current persistence methods plus a new
  `update_log_counts` to eliminate the last `db.query()` from the
  application layer
- `app/versions/adapters/repository.py` — `SqlAlchemyVersionRepository`
  concrete implementation (lifted from the legacy module)
- `app/versions/application/service.py` — `VersionUpdateService`
  refactored to receive the `VersionRepository` Port via constructor
  injection; no longer takes a raw `Session`
- `app/versions/api/dependencies.py` — DI seam that binds
  `SqlAlchemyVersionRepository` to the Port (the only file under `api/`
  permitted to import from `adapters/` per ARCHITECTURE.md §4.4)
- `tests/unit/versions/fakes.py` — `FakeVersionRepository`, an
  in-memory dict-backed implementation for unit tests
- `tests/unit/versions/test_service_with_fake.py` — protocol-conformance
  checks plus Fake-driven service tests demonstrating the target style

Backward-compat shims at the historical paths (`app/versions/repository.py`,
`app/versions/service.py`) preserve existing imports and the legacy
`VersionUpdateService(db: Session)` constructor while the rest of the
codebase migrates. They are documented as deprecated and will be removed
once consumers move to `get_version_service`.

Test patches that referenced `app.versions.service.minecraft_version_manager`
are updated to point at the new application-layer module.

Refs: #154 / #221
