+++ app/versions/adapters/__init__.py
@@ -1,0 +1,6 @@
+"""Adapter implementations for the versions domain Ports.
+
+These modules concretely implement the Protocols defined in
+`app/versions/domain/ports.py`. They may depend on SQLAlchemy and other
+frameworks. They must not be imported from `domain/` or `application/`.
+"""
