+++ app/versions/api/__init__.py
@@ -1,0 +1,5 @@
+"""API layer for the versions module.
+
+FastAPI routers and DTOs. `dependencies.py` is the only file in this layer
+allowed to import from `adapters/`.
+"""
