+++ app/versions/application/__init__.py
@@ -1,0 +1,5 @@
+"""Application layer for the versions module.
+
+Use cases that depend only on Ports defined in `domain/`. Must not import
+from `adapters/` or `api/`.
+"""
