+++ app/versions/domain/__init__.py
@@ -1,0 +1,5 @@
+"""Domain layer for the versions module.
+
+Contains Port (Protocol) definitions and pure domain logic.
+Must not import any framework, database driver, or HTTP client.
+"""
